Depreciate CpcQualityMeasureIdValidator and update PcfQualityMeasureIdValidator
